requirements.txt
fastapi
pydantic
uvicorn[standard]
mysql-connector-python
pandas
python-dotenv
python-multipart
requests